# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "poprako-server",
    "description": "Event-driven backend service for manga translation project management",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/chapters/{chapter_id}": {
      "get": {
        "tags": [
          "chapters"
        ],
        "summary": "`GET /api/v1/chapters/{chapter_id}` — fetch a chapter by id.",
        "operationId": "get_info",
        "parameters": [
          {
            "name": "chapter_id",
            "in": "path",
            "description": "Chapter ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chapter info retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpBody_ChapterInfoVal"
                }
              }
            }
          },
          "403": {
            "description": "No permission to view this chapter"
          },
          "404": {
            "description": "Chapter not found"
          }
        }
      },
      "delete": {
        "tags": [
          "chapters"
        ],
        "summary": "`DELETE /api/v1/chapters/{chapter_id}` — delete a chapter and descendants.",
        "operationId": "delete",
        "parameters": [
          {
            "name": "chapter_id",
            "in": "path",
            "description": "Chapter ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Chapter deleted"
          },
          "403": {
            "description": "No permission to delete this chapter"
          },
          "404": {
            "description": "Chapter not found"
          }
        }
      },
      "patch": {
        "tags": [
          "chapters"
        ],
        "summary": "`PATCH /api/v1/chapters/{chapter_id}` — partially update a chapter's profile.",
        "operationId": "update_info",
        "parameters": [
          {
            "name": "chapter_id",
            "in": "path",
            "description": "Chapter ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChapterInfoParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Chapter updated"
          },
          "403": {
            "description": "No permission to update this chapter"
          },
          "404": {
            "description": "Chapter not found"
          },
          "422": {
            "description": "Path id does not match body id"
          }
        }
      }
    },
...
    "/api/v1/comics/{comic_id}": {
      "get": {
        "tags": [
          "comics"
        ],
        "summary": "`GET /api/v1/comics/{comic_id}` — fetch a comic by id.",
        "operationId": "get_info",
        "parameters": [
          {
            "name": "comic_id",
            "in": "path",
            "description": "Comic ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comic info retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpBody_ComicInfoVal"
                }
              }
            }
          },
          "403": {
            "description": "No permission to view this comic"
          },
          "404": {
            "description": "Comic not found"
          }
        }
      },
      "put": {
        "tags": [
          "comics"
        ],
        "summary": "`PUT /api/v1/comics/{comic_id}` — update a comic's profile.",
        "operationId": "update_info",
        "parameters": [
          {
            "name": "comic_id",
            "in": "path",
            "description": "Comic ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateComicInfoParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Comic updated"
          },
          "403": {
            "description": "No permission to update this comic"
          },
          "404": {
            "description": "Comic not found"
          },
          "422": {
            "description": "Path id does not match body id"
          }
        }
      },
      "delete": {
        "tags": [
          "comics"
        ],
        "summary": "`DELETE /api/v1/comics/{comic_id}` — delete a comic and descendants.",
        "operationId": "delete",
        "parameters": [
          {
            "name": "comic_id",
            "in": "path",
            "description": "Comic ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Comic deleted"
          },
          "403": {
            "description": "No permission to delete this comic"
          },
          "404": {
            "description": "Comic not found"
          }
        }
      }
    },
...
    "/api/v1/teams/{team_id}": {
      "get": {
        "tags": [
          "teams"
        ],
        "summary": "`GET /api/v1/teams/{team_id}` — fetch a team by id.",
        "operationId": "get_info",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "description": "Team ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Team info retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpBody_TeamInfoVal"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Team not found"
          }
        }
      },
      "put": {
        "tags": [
          "teams"
        ],
        "summary": "`PUT /api/v1/teams/{team_id}` — update a team's profile.",
        "operationId": "update_info",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "description": "Team ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeamInfoParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Team updated"
          },
          "403": {
            "description": "No permission to update this team"
          },
          "404": {
            "description": "Team not found"
          },
          "422": {
            "description": "Path id does not match body id"
          }
        }
      },
      "delete": {
        "tags": [
          "teams"
        ],
        "summary": "`DELETE /api/v1/teams/{team_id}` — delete a team and all descendants.",
        "operationId": "delete",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "description": "Team ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Team deleted"
          },
          "403": {
            "description": "No permission to delete this team"
          },
          "404": {
            "description": "Team not found"
          }
        }
      }
    },
...
    "/api/v1/teams/{team_id}/avatar/mark-uploaded": {
      "post": {
        "tags": [
          "teams"
        ],
        "summary": "`POST /api/v1/teams/{team_id}/avatar/mark-uploaded` — confirm a team avatar upload.",
        "operationId": "mark_avatar_uploaded",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "description": "Team ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkTeamAvatarUploadedParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Avatar upload confirmed"
          },
          "403": {
            "description": "No permission to modify this team's avatar"
          },
          "404": {
            "description": "Team not found"
          }
        }
      }
    },
    "/api/v1/teams/{team_id}/avatar/reserve": {
      "post": {
        "tags": [
          "teams"
        ],
        "summary": "`POST /api/v1/teams/{team_id}/avatar/reserve` — reserve a team avatar upload slot.",
        "operationId": "reserve_avatar",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "description": "Team ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveTeamAvatarParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Avatar upload URL reserved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpBody_ReserveTeamAvatarPayload"
                }
              }
            }
          },
          "403": {
            "description": "No permission to modify this team's avatar"
          },
          "404": {
            "description": "Team not found"
          }
        }
      }
    },
...
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "`GET /api/v1/users/me` — current user's profile.",
        "operationId": "get_my_info",
        "responses": {
          "200": {
            "description": "Current user profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpBody_UserInfoVal"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          }
        }
      }
    },
    "/api/v1/users/{user_id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "`GET /api/v1/users/{user_id}` — a user's profile by id.",
        "operationId": "get_info",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "Target user ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User profile retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpBody_UserInfoVal"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "tags": [
          "users"
        ],
        "summary": "`PUT /api/v1/users/{user_id}` — update a user's profile.",
        "operationId": "update_info",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "Target user ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserInfoParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Profile updated"
          },
          "403": {
            "description": "Cannot modify another user's profile"
          },
          "409": {
            "description": "QID already taken"
          },
          "422": {
            "description": "Path id does not match body id"
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "`DELETE /api/v1/users/{user_id}` — delete a user account.",
        "operationId": "delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "Target user ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted"
          },
          "403": {
            "description": "Cannot delete another user's account"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/api/v1/users/{user_id}/avatar/mark-uploaded": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "`POST /api/v1/users/{user_id}/avatar/mark-uploaded` — confirm an avatar upload.",
        "operationId": "mark_avatar_uploaded",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "Target user ID (must match authenticated user)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkUserAvatarUploadedParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Avatar upload confirmed"
          },
          "403": {
            "description": "Cannot confirm another user's avatar"
          }
        }
      }
    },
    "/api/v1/users/{user_id}/avatar/reserve": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "`POST /api/v1/users/{user_id}/avatar/reserve` — reserve an avatar upload slot.",
        "operationId": "reserve_avatar",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "Target user ID (must match authenticated user)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveUserAvatarParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Avatar upload URL reserved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpBody_ReserveUserAvatarPayload"
                }
              }
            }
          },
          "403": {
            "description": "Cannot modify another user's avatar"
          }
        }
      }
    },
...
    "/api/v1/worksets/{workset_id}": {
      "get": {
        "tags": [
          "worksets"
        ],
        "summary": "`GET /api/v1/worksets/{workset_id}` — fetch a workset by id.",
        "operationId": "get_info",
        "parameters": [
          {
            "name": "workset_id",
            "in": "path",
            "description": "Workset ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workset info retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpBody_WorksetInfoVal"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "No permission to view this workset"
          },
          "404": {
            "description": "Workset not found"
          }
        }
      },
      "put": {
        "tags": [
          "worksets"
        ],
        "summary": "`PUT /api/v1/worksets/{workset_id}` — update a workset's profile.",
        "operationId": "update_info",
        "parameters": [
          {
            "name": "workset_id",
            "in": "path",
            "description": "Workset ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorksetInfoParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Workset updated"
          },
          "403": {
            "description": "No permission to update this workset"
          },
          "404": {
            "description": "Workset not found"
          },
          "422": {
            "description": "Path id does not match body id"
          }
        }
      },
      "delete": {
        "tags": [
          "worksets"
        ],
        "summary": "`DELETE /api/v1/worksets/{workset_id}` — delete a workset and descendants.",
        "operationId": "delete",
        "parameters": [
          {
            "name": "workset_id",
            "in": "path",
            "description": "Workset ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Workset deleted"
          },
          "403": {
            "description": "No permission to delete this workset"
          },
          "404": {
            "description": "Workset not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HttpBody_ChapterInfoVal": {
        "type": "object",
        "description": "Success response body wrapping a usecase value.\n\nSerializes as the standard JSON success envelope. HTTP metadata such as\nstatus code, extra headers, and `Set-Cookie` values are not part of the JSON\nbody.",
        "required": [
          "code",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "example": 0,
            "minimum": 0
          },
          "data": {
            "type": "object",
            "description": "Presentation-ready chapter information.\n\nMirrors [`ChapterInfo`] but converts timestamps to Unix milliseconds\nand exposes the same grouped field layout as the API response.\n\nConstruct via [`From<ChapterInfo>`] — the conversion is infallible.\n\n[`ChapterInfo`]: crate::model::chapter::ChapterInfo",
            "required": [
              "id",
              "comic_id",
              "is_pinned",
              "index",
              "subtitle",
              "page_count",
              "total_unit_count",
              "translated_unit_count",
              "proofread_unit_count",
              "stages",
              "creator_id",
              "created_at",
              "updated_at"
            ],
            "properties": {
              "comic": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/ComicInfoVal"
                  }
                ]
              },
              "comic_id": {
                "type": "string"
              },
              "created_at": {
                "type": "integer",
                "format": "int64"
              },
              "creator": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/UserInfoVal"
                  }
                ]
              },
              "creator_id": {
                "type": "string"
              },
              "id": {
                "type": "string"
              },
              "index": {
                "type": "integer",
                "format": "int32"
              },
              "is_pinned": {
                "type": "boolean"
              },
              "page_count": {
                "type": "integer",
                "format": "int32"
              },
              "proofread_unit_count": {
                "type": "integer",
                "format": "int32"
              },
              "stages": {
                "$ref": "#/components/schemas/StageMask"
              },
              "subtitle": {
                "type": "string"
              },
              "total_unit_count": {
                "type": "integer",
                "format": "int32"
              },
              "translated_unit_count": {
                "type": "integer",
                "format": "int32"
              },
              "updated_at": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        }
      },
      "HttpBody_ComicInfoVal": {
        "type": "object",
        "description": "Success response body wrapping a usecase value.\n\nSerializes as the standard JSON success envelope. HTTP metadata such as\nstatus code, extra headers, and `Set-Cookie` values are not part of the JSON\nbody.",
        "required": [
          "code",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "example": 0,
            "minimum": 0
          },
          "data": {
            "type": "object",
            "description": "Presentation-ready comic information.\n\nMirrors [`ComicInfo`] but converts timestamps to Unix milliseconds and\nresolves the cover key to a signed download URL via [`ImagePool`] when\nthe cover has been uploaded.\n\nConstruct via [`ComicInfoVal::from_model`] — the conversion requires\nan [`ImagePool`] instance for URL signing.\n\n[`ComicInfo`]: crate::model::comic::ComicInfo",
            "required": [
              "id",
              "workset_id",
              "index",
              "title",
              "author",
              "chapter_count",
              "chapter_next_index",
              "creator_id",
              "last_active_at",
              "created_at",
              "updated_at"
            ],
            "properties": {
              "author": {
                "type": "string"
              },
              "chapter_count": {
                "type": "integer",
                "format": "int32"
              },
              "chapter_next_index": {
                "type": "integer",
                "format": "int32"
              },
              "cover_url": {
                "type": [
                  "string",
                  "null"
                ],
                "description": "Resolved signed download URL for the cover image, or [`None`] if\nno cover has been uploaded."
              },
              "created_at": {
                "type": "integer",
                "format": "int64"
              },
              "creator": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/UserInfoVal"
                  }
                ]
              },
              "creator_id": {
                "type": "string"
              },
              "description": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "id": {
                "type": "string"
              },
              "index": {
                "type": "integer",
                "format": "int32"
              },
              "last_active_at": {
                "type": "integer",
                "format": "int64"
              },
              "pinned_chapter": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/ChapterInfoVal"
                  }
                ]
              },
              "team": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/TeamInfoVal"
                  }
                ]
              },
              "title": {
                "type": "string"
              },
              "updated_at": {
                "type": "integer",
                "format": "int64"
              },
              "workset": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/WorksetInfoVal"
                  }
                ]
              },
              "workset_id": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "MarkTeamAvatarUploadedParams": {
        "type": "object",
        "description": "Input parameters for confirming a team avatar upload completed.",
        "required": [
          "avatar_version"
        ],
        "properties": {
          "avatar_version": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "ReserveTeamAvatarParams": {
        "type": "object",
        "description": "Input parameters for reserving a new team avatar upload slot.",
        "required": [
          "file_ext"
        ],
        "properties": {
          "file_ext": {
            "type": "string"
          }
        }
      },
...
      "ReserveUserAvatarParams": {
        "type": "object",
        "description": "Input parameters for reserving a new avatar upload slot.",
        "required": [
          "file_ext"
        ],
        "properties": {
          "file_ext": {
            "type": "string"
          }
        }
      },
...
      "UpdateChapterInfoParams": {
        "type": "object",
        "description": "Input parameters for partially updating a chapter's profile.",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "pin": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "subtitle": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "UpdateComicInfoParams": {
        "type": "object",
        "description": "Input parameters for updating a comic's title, author, and description.\n\nCover updates are handled by dedicated endpoints.\n\n[`reserve_cover`]: crate::usecase::comic::reserve_cover",
        "required": [
          "id",
          "title",
          "author"
        ],
        "properties": {
          "author": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "UpdateTeamInfoParams": {
        "type": "object",
        "description": "Input parameters for updating a team's profile.",
        "required": [
          "id",
          "name",
          "description"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}